+++ CHANGELOG.md
@@ -34,6 +34,8 @@ and this project adheres to
   - [#1514](https://github.com/iovisor/bpftrace/pull/1514)
 - Support for tracepoint __data_loc fields
   - [#1542](https://github.com/iovisor/bpftrace/pull/1542)
+- Set addrspace info for various builtins
+  - [#1504](https://github.com/iovisor/bpftrace/pull/1504)
 
 #### Changed
 - Warn if using `print` on `stats` maps with top and div arguments